improve & rename partitionRandomly()
Renamed to partition(). Length of the returned array object is now always the integer specified in the second argument.  If the absolute value of the ratio obtained by dividing the first argument by the second argument is less than one, logs an error message to the console instead of getting stuck in an infinite loop.